+++ lib/src/main/java/com/github/rvesse/airline/model/CommandGroupMetadata.java
@@ -6,63 +6,61 @@ import com.google.common.base.Function;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 
-public class CommandGroupMetadata
-{
+public class CommandGroupMetadata {
     private final String name;
     private final String description;
+    private final boolean hidden;
     private final List<OptionMetadata> options;
     private final CommandMetadata defaultCommand;
     private final List<CommandMetadata> commands;
 
-    public CommandGroupMetadata(String name, String description, Iterable<OptionMetadata> options, CommandMetadata defaultCommand, Iterable<CommandMetadata> commands)
-    {
+    public CommandGroupMetadata(String name, String description, boolean hidden, Iterable<OptionMetadata> options,
+            CommandMetadata defaultCommand, Iterable<CommandMetadata> commands) {
         this.name = name;
         this.description = description;
+        this.hidden = hidden;
         this.options = ImmutableList.copyOf(options);
         this.defaultCommand = defaultCommand;
         this.commands = Lists.newArrayList(commands);
     }
 
-    public String getName()
-    {
+    public String getName() {
         return name;
     }
 
-    public String getDescription()
-    {
+    public String getDescription() {
         return description;
     }
 
-    public List<OptionMetadata> getOptions()
-    {
+    public boolean isHidden() {
+        return hidden;
+    }
+
+    public List<OptionMetadata> getOptions() {
         return options;
     }
 
-    public CommandMetadata getDefaultCommand()
-    {
+    public CommandMetadata getDefaultCommand() {
         return defaultCommand;
     }
 
-    public List<CommandMetadata> getCommands()
-    {
+    public List<CommandMetadata> getCommands() {
         return ImmutableList.copyOf(commands);
     }
-    
-    public void addCommand(CommandMetadata command)
-    {
-        if(!commands.contains(command))
-        {
+
+    public void addCommand(CommandMetadata command) {
+        if (!commands.contains(command)) {
             commands.add(command);
         }
     }
 
     @Override
-    public String toString()
-    {
+    public String toString() {
         final StringBuilder sb = new StringBuilder();
         sb.append("CommandGroupMetadata");
         sb.append("{name='").append(name).append('\'');
         sb.append(", description='").append(description).append('\'');
+        sb.append(", hidden=").append(hidden);
         sb.append(", options=").append(options);
         sb.append(", defaultCommand=").append(defaultCommand);
         sb.append(", commands=").append(commands);
@@ -70,12 +68,9 @@ public class CommandGroupMetadata
         return sb.toString();
     }
 
-    public static Function<CommandGroupMetadata, String> nameGetter()
-    {
-        return new Function<CommandGroupMetadata, String>()
-        {
-            public String apply(CommandGroupMetadata input)
-            {
+    public static Function<CommandGroupMetadata, String> nameGetter() {
+        return new Function<CommandGroupMetadata, String>() {
+            public String apply(CommandGroupMetadata input) {
                 return input.getName();
             }
         };