+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # Airline - Change Log
 
+## 1.0.2
+
+
+
+## 1.0.1
+
+- Fix regression in `RonnCommandUsageGenerator`
+
 ## 1.0.0
 
 - Code Structure Refactoring
     - Root package is now `com.github.rvesse.airline`
     - Main library now lives under `lib/` in source control
     - New examples module under `examples/` in source control
+- Annotation Improvements
+    - `@Group` can now be marked as `hidden`
+    - `discussion` parameter of `@Command` is now a `String[]` rather than a single string making it easier to specify long descriptions
 - CLI Improvements
     - User defined command aliases are now supported
     - CLI builder classes are now public and have their own `builder` package
@@ -19,11 +30,20 @@
     - Bash completion fixes and improvements:
         - Default command and default group command completion now included
         - Functions for default group commands (i.e. top level commands) are no longer missing if there are also groups present
+    - Fixed a bug with hidden options not displaying in synopsis even when including hidden options was enabled
     - `HelpOption` improved:
         - Generated help will include program and group name where applicable
         - It can now be used to show help with an arbitrary usage generator
         - `showHelpIfRequested()` guarantees to only display help once
         - `showHelp()` can be used to display help regardless
+    - Command usage generators now print each item in the discussion as a separate paragraph
+- IO Improvements
+    - New `com.github.rvesse.airline.io` package with useful helper stuff for doing advanced console IO
+    - Support for colorised output streams and writers:
+        - Basic ANSI Colors (8 Colors, normal and bright variants)
+        - 256 Colors (Basic ANSI Colors plus 3 colour palettes plus grayscale palette) - See [color chart](https://camo.githubusercontent.com/6378594a85c578517c5a4e494789bd4d66c9e46b/68747470733a2f2f7261772e6769746875622e636f6d2f666f697a652f676f2e7367722f6d61737465722f787465726d5f636f6c6f725f63686172742e706e67) for more detail
+        - True Color (24 bit colour i.e. 16 million colours) - Many terminals may not support this mode
+        - Includes basic text decorations e.g. bold, underline
 
 ## 0.9.2